Fix bug with flags
Flag did not reset when falling into the void with it in the player's inventory. This bug is fixed, plus a couple of small quality of life updates.

Gamemodes get set to the proper mode when building or starting a game.

The void damage event now gets cancelled to avoid the damage tick that occurs at the same time as teleportation.